# Support multi-thread model weight loading