<!-- Thank you for your contribution! We appreciate it. The following guidelines will help improve your pull request and facilitate feedback. If anything is unclear, don't hesitate to submit your pull request and ask the maintainers for assistance. -->

## Motivation
Support multi-thread model weight loading when using the --load-format "auto", which can speedup weight load process for users with distributed storage backend or NVME ssd disk. 

Tests have found that storing DeepSeek-R1-0528 model weights on a single NVME drive for cold start of services can speed up model loading time by 6.26 times


<!-- Explain the purpose of this PR and the goals it aims to achieve. -->

## Modifications
base on PR https://github.com/sgl-project/sglang/issues/4822

- Multi-threaded Weight Loading: Implemented multi-threaded loading for safetensors and PyTorch (.pt/.bin) model weights, significantly accelerating the model loading process, especially for large sharded models.

- Introduced a new --model-loader-extra-config command-line argument: Allowing users to enable multi-threaded loading (enable_multithread_load) and specify the number of threads (num_threads) via a JSON string.

 
Cmd Example: 
```bash
python3 -m sglang.launch_server \
 --model-path /data/DeepSeek-R1-0528 \
 --tp 8 \
 --trust-remote-code \
 --mem-fraction-static 0.9 \
 --host 0.0.0.0 \
 --port 8001 \
 --enable-metric \
 --log-requests \
 --log-requests-level 0 \
 --allow-auto-truncate  \
 --quantization fp8 \
 --model-loader-extra-config '{"enable_multithread_load": true, "num_threads": 8}'
```


Test DeepSeek-R1-0528 safetensor loading in NVME ssd PCIE Gen4 (Speed 16GT/s, Width x4)
|                                    | Time consumption before optimization | Time consumption after optimization | Load Weight Speed up |
| ---------------------------------- | ------------------------------------ | ----------------------------------- | -------------------- |
| weight in single NVME ssd          | 702s                                 | 112s                                | 6.26x                |
| weight in raid0 with four NVME ssd | 656s                                 | 62s                                 | 10.58x               |



<!-- Describe the changes made in this PR. -->

## Checklist


- [x] Format your code according to the [Code Formatting with Pre-Commit](https://docs.sglang.ai/references/contribution_guide.html#code-formatting-with-pre-commit).
- [ ] Add unit tests as outlined in the [Running Unit Tests](https://docs.sglang.ai/references/contribution_guide.html#running-unit-tests-adding-to-ci).
- [ ] Update documentation / docstrings / example tutorials as needed, according to [Writing Documentation](https://docs.sglang.ai/references/contribution_guide.html#writing-documentation-running-docs-ci).
- [ ] Provide throughput / latency benchmark results and accuracy evaluation results as needed, according to [Benchmark and Profiling](https://docs.sglang.ai/references/benchmark_and_profiling.html) and [Accuracy Results](https://docs.sglang.ai/references/accuracy_evaluation.html).
- [ ] For reviewers: If you haven't made any contributions to this PR and are only assisting with merging the main branch, please remove yourself as a co-author when merging the PR.
- [ ] Please feel free to join our Slack channel at https://slack.sglang.ai to discuss your PR.- [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ] - [ ]