@xianzhiT 2025-06-18
> Disclaimer: I'm not a core contributor - my review was requested by @zhaochenyang20
> 
> Thanks for the contribution! The implementation looks good overall, though there are a few minor style points that could be addressed.
> 
> One question: does the author report the total launch time, or specifically the model loading time (disk to memory)? I'm curious about the multi-threading performance gains for end-to-end launch time from disk to GPU, even though PCIe bandwidth should theoretically be much higher.

Thanks for the review.

> Disclaimer: I'm not a core contributor - my review was requested by @zhaochenyang20
> 
> Thanks for the contribution! The implementation looks good overall, though there are a few minor style points that could be addressed.
> 
> One question: does the author report the total launch time, or specifically the model loading time (disk to memory)? I'm curious about the multi-threading performance gains for end-to-end launch time from disk to GPU, even though PCIe bandwidth should theoretically be much higher.