@BraveY 2025-06-18
@xianzhiT , @garrett4wade  — Great work! I submitted a similar PR a few days ago, aiming to mmap the weight file into pagecache to mitigate cold start issues in inference. Would either of you be willing to review my PR and share your feedback?