@xianzhiT 2025-06-18
> Disclaimer: I'm not a core contributor - my review was requested by @zhaochenyang20
> 
> Thanks for the contribution! The implementation looks good overall, though there are a few minor style points that could be addressed.
> 
> One question: does the author report the total launch time, or specifically the model loading time (disk to memory)? I'm curious about the multi-threading performance gains for end-to-end launch time from disk to GPU, even though PCIe bandwidth should theoretically be much higher.



> > Disclaimer: I'm not a core contributor - my review was requested by @zhaochenyang20
> > Thanks for the contribution! The implementation looks good overall, though there are a few minor style points that could be addressed.
> > One question: does the author report the total launch time, or specifically the model loading time (disk to memory)? I'm curious about the multi-threading performance gains for end-to-end launch time from disk to GPU, even though PCIe bandwidth should theoretically be much higher.
> 
> Thanks for the review, appreciate your feedback. I'll address the minor style points you mentioned and update the code accordingly.
> 
> The report is the time consume for model weight loading from disk to memory. I calculated the time by counting the time interval between "Load weight begin" and "Load weight end" in the log.

The nvme ssd disk I use to store model weights is PCIE GEN4 width 4, the theoretical upper limit for read bandwidth is 8GB/s. In the optimized test, I observed through iotop that the read bandwidth can reach around 6.3GB, which is close to the theoretical bandwidth limit.
<img width="678" alt="Clipboard_Screenshot_1750218743" src="https://github.com/user-attachments/assets/22c90d7c-e449-415a-89ae-738e5cc0954e" />

I have also test the e2e launch time speed, we can get server cold start in about 160s.

|                           | E2E Launch Time consumption before optimization (DeepGEMM JIT cached) | E2E Launch Time consumption after optimization ((DeepGEMM JIT cached)) | Speedup |
| ------------------------- | ------------------------------------------------------------ | ------------------------------------------------------------ | ------- |
| weight in single NVME ssd | 756s                                                         | 160s                                                         | 4.725x  |